creating llm wrapper
