+++ src/scraper.py
@@ -1,146 +1,91 @@
-# src/storage.py
+# src/scraper.py
 import json
 import logging
 from pathlib import Path
-from abc import ABC, abstractmethod
-import re
 from typing import List, Dict, Any
-from google.cloud import firestore
 
-_RE_NON_ALNUM = re.compile(r"[^a-z0-9]+")
-_RE_UNDERSCORES = re.compile(r"_+")
+from crawl4ai import (
+    AsyncWebCrawler,
+    BrowserConfig,
+    CrawlerRunConfig,
+    CacheMode,
+    JsonCssExtractionStrategy,
+    LXMLWebScrapingStrategy,
+)
 
-class StorageBackend(ABC):
-    @abstractmethod
-    async def get_urls(self, source: str) -> List[str]: ...
-    @abstractmethod
-    async def save_urls(self, source: str, urls: List[str]) -> None: ...
+from src.config import SourceConfig
 
-    @abstractmethod
-    async def get_schema(self, source: str) -> Dict[str, Any]: ...
-    @abstractmethod
-    async def save_schema(self, source: str, schema: Dict[str, Any]) -> None: ...
+async def scrape_urls(
+    urls: List[str],
+    schema: Dict[str, Any],
+    source: SourceConfig
+) -> List[Dict[str, Any]]:
+    return await _scrape_with_schema(
+        urls=urls,
+        schema=schema,
+        max_concurrency=source.max_concurrency
+    )
 
-    @abstractmethod
-    async def get_data(self, source: str) -> List[Dict[str, Any]]: ...
-    @abstractmethod
-    async def save_data(self, source: str, data: List[Dict[str, Any]]) -> None: ...
+async def _scrape_with_schema(
+    urls: List[str],
+    schema: Dict[str, Any],
+    max_concurrency: int = 10,
+) -> List[Dict[str, Any]]:
+    """
+    Apply the JSON-CSS schema to each URL in parallel using arun_many.
+    Returns a flat list of all extracted course dicts, each with a "_source_url".
+    """
+    # 1) Setup crawler config
+    browser_cfg = BrowserConfig(headless=True, verbose=False)
 
-class FirestoreStorage(StorageBackend):
-    def __init__(self, project=None):
-        self.db = firestore.AsyncClient(project=project)
-        self.root = self.db.collection("scraper")
-        self.log = logging.getLogger(__name__)
+    extraction_strategy = JsonCssExtractionStrategy(schema)
+
+    run_cfg = CrawlerRunConfig(
+        cache_mode=CacheMode.BYPASS,
+        scraping_strategy=LXMLWebScrapingStrategy(),
+        extraction_strategy=extraction_strategy,
+    )
+
+    all_records: List[Dict[str, Any]] = []
+
+    # 2) Fire off all URLs in parallel
+    async with AsyncWebCrawler(config=browser_cfg) as crawler:
+        results = await crawler.arun_many(
+            urls=urls,
+            config=run_cfg,
+            max_concurrency=max_concurrency
+        )
     
-    def _slugify(self, s: str) -> str:
-        # Lowercase, replace non-alnum with underscore, collapse multiple underscores
-        s = s.lower()
-        s = _RE_NON_ALNUM.sub("_", s)
-        s = _RE_UNDERSCORES.sub("_", s).strip("_")
-        return s or "no_title"
+    # 3) Parse out each page's JSON payload
+    for page_result in results:
+        raw = page_result.extracted_content
+        try:
+            items = json.loads(raw)
+        except json.JSONDecodeError:
+            continue
 
-    async def get_urls(self, source_name: str) -> List[str]:
-        doc = self.root.document(source_name).collection("urls").document("list")
-        snap = await doc.get()
-        self.log.info(f"Loaded URLs for {source_name}")
-        return (snap.to_dict() or {}).get("items", [])
-
-    async def save_urls(self, source_name: str, urls: List[str]) -> None:
-        doc = self.root.document(source_name).collection("urls").document("list")
-        self.log.info(f"Saving {len(urls)} URLs to {source_name}")
-        await doc.set({"items": urls})
-
-    async def get_schema(self, source_name: str) -> Dict[str, Any]:
-        doc = self.root.document(source_name).collection("schema").document("definition")
-        snap = await doc.get()
-        self.log.info(f"Loaded schema for {source_name}")
-        return snap.to_dict() or {}
-
-    async def save_schema(self, source_name: str, schema: Dict[str, Any]) -> None:
-        doc = self.root.document(source_name).collection("schema").document("definition")
-        await doc.set(schema)
-        self.log.info(f"Saved schema to {source_name}")
-
-    async def get_data(self, source_name: str) -> List[Dict[str, Any]]:
-        col = self.root.document(source_name).collection("courses")
-        snapshots = await col.get()
-        self.log.info(f"Loaded {len(snapshots)} course documents for {source_name}")
-        return [doc.to_dict() for doc in snapshots]
-
-    async def save_data(self, source_name: str, data: List[Dict[str, Any]]) -> None:
-        MAX_BATCH_WRITE = 500
-
-        col   = self.root.document(source_name).collection("courses")
-        self.log.info(f"Saving {len(data)} course records to {source_name} in chunks of {MAX_BATCH_WRITE}")
-        for i in range(0, len(data), MAX_BATCH_WRITE):
-            batch = self.db.batch()
-            chunk = data[i:i + MAX_BATCH_WRITE]
-            for course in chunk:
-                raw_code = course.get("course_code") # Attempt to pull course_code for ID
-                if isinstance(raw_code, str) and raw_code.strip():
-                    course_id = raw_code.strip()
+        source_url = getattr(page_result, "url", None) or getattr(page_result, "request_url", None)
+        for item in items:
+            item["_source_url"] = source_url
+            if "course_code" in item and isinstance(item["course_code"], list) and item["course_code"]:
+                str_codes: list[str] = []
+                raw_codes = item.get("course_code")
+                for code in raw_codes:
+                    if isinstance(code, dict):
+                        txt = str(code.get("text", "")).strip()
+                    else:
+                        txt = str(code).strip()
+                    if txt:
+                        str_codes.append(txt)
+                if str_codes:
+                    unique = sorted(set(str_codes))
+                    norm = "_".join(unique)
+                    item["course_code"] = norm
+                    log = logging.getLogger(__name__)
+                    log.debug(f"Codes {raw_codes!r} → {norm!r}")
                 else:
-                    raw_title = course.get("course_title", "") # Fall back to title if no valid course_code
-                    course_id = self._slugify(raw_title)
+                    item.pop("course_code", None)
 
-                course_id = self._slugify(course_id)
-                ref = col.document(course_id)
-                batch.set(ref, course)  # overwrite if it already exists
+            all_records.append(item)
 
-            self.log.info(f"Saving {len(batch)} course records to {source_name}")
-            await batch.commit()
-
-class LocalFileStorage(StorageBackend):
-    def __init__(self, base_dir: Path):
-        self.base_dir = base_dir
-        self.base_dir.mkdir(parents=True, exist_ok=True)
-
-    def _ensure_dir(self, source_name: str) -> Path:
-        out = self.base_dir / source_name
-        out.mkdir(exist_ok=True)
-        return out
-    
-    async def get_urls(self, source_name: str) -> List[str]:
-        d = self.base_dir / source_name
-        if not d.exists():
-            return []
-        try:
-            with open(d / "urls.json", "r") as f:
-                return json.load(f)
-        except:
-            return []
-
-    async def save_urls(self, source_name: str, urls: List[str]) -> None:
-        d = self._ensure_dir(source_name)
-        with open(d / "urls.json", "w") as f:
-            json.dump(list(urls), f, indent=2)
-
-    async def save_schema(self, source_name: str, schema: Dict[str, Any]) -> None:
-        d = self._ensure_dir(source_name)
-        with open(d / "schema.json", "w") as f:
-            json.dump(schema, f, indent=2)
-
-    async def get_schema(self, source_name: str) -> Dict[str, Any]:
-        d = self.base_dir / source_name
-        if not d.exists():
-            return {}
-        try:
-            with open(d / "schema.json", "r") as f:
-                return json.load(f)
-        except FileNotFoundError:
-            return {}
-
-    async def save_data(self, source_name: str, data: List[Dict[str, Any]]) -> None:
-        d = self._ensure_dir(source_name)
-        with open(d / "courses.json", "w") as f:
-            json.dump(list(data), f, indent=2)
-
-    async def get_data(self, source_name: str) -> List[Dict[str, Any]]:
-        d = self.base_dir / source_name
-        if not d.exists():
-            return []
-        try:
-            with open(d / "courses.json", "r") as f:
-                return json.load(f)
-        except FileNotFoundError:
-            return []
+    return all_records
